chore: reformat testing files
